Remove frontend and clean up
We're splitting the Alpha app in to a backend and frontend service while
maintaining the majority of the code written in alpha to give us a
headstart on implementing the beta service.  This PR removes all
frontend view and components, cleans up the package list and updates the
read me.