Fix bugs in the Upload API point
Through the various iterations of Alpha the API layer and frontend
layers stopped working in the same way at the point of file upload.  Now
the frontend is entirely dependant on the backend it was important to
fix the backend to behave as the old alpha frontend had.

This should fix issues with uploading files using the API layer not
having IDs, being recorded in the Database and not showing up in the
file list.