SW-138 Update backend for sources and dimenions
This updates the back with 2 new routes.

The is for confirming the uploaded fact table and moving it from
tempoary blob storage to tghe datalate.

The second route handles labelling sources as Dimension, Footnotes and
DataValues as well as the option to ignore a source entirly.

Also include is some bug fixes within the Data model and database
script.