Rationalised all tests, made all UUIDs lowercase
Rationalised the tests following the rebase and reorganised them so they
are easier to manage and follow.

Following a conversation with @wheelsandcogs the otherday we realised
that PostgreSQL stores UUIDs in lowercase which might cause a problem
when asking the database with an uppercase UUID.  So I've appeneded a
.toLowerCase() whenever we get a UUID as a string or generate a random
UUID using code.