Updated the backend to process CSV straight to SQLite3
This updates the backend to copy the csv data directly to an SQLite3
file which is much easier to query and much faster.  This allows us to
get SQLite3 to handle things like pagination, pulling out small parts of
the facts table or single columns as needed for the source preview.