Move internal CSV parsing to use DuckDB.
This moves the internal architecture of the app from parsing CSVs
manually to loading them in to DuckDB and running proper SQL type
queries against the CSV file.  This is significantly faster and results
in significantly less code.

Other changes include abandoning the use of Azure BlobStorage as a
temporary staging area for user uploaded CSVs and moving back to a
single blob store in the shape of Azure DataLake.

Data in the Data-Lake is now stored as `/<store>/<dataset-id>/<file-id>`
to organise the Data-Lake more sensibly.