Change the datamodel to something we have more confidence in
Following a number of discussions and attempting to implementing a cube
in code it became clear the datamodel we had previously been working
with wasn't going to allow us to successfully produce any sort of cube
in time for the upcoming MVP deadline.

Its hoped this new simpler datamodel will allow us to accellerate
development slightly and allow for the production of a working cube in
working software.