The start of a big refactor, added in measure handling
The intention of this commit is to start the process of trying to reduce
some of the code duplication and to make the code a little easier to
read and maintain.  The cube handler has be refactored in to smaller
functions.  Exported interfaces and functions which are used in a number
of places have been moved out in to files of their own and there's now
an extractors folder for the extractors used by dimensions and measure.

Also added is the measure handling code to allow the uploading of
measures.  The intention is to proper measure handling and note code
handling in to the cube handler.