Fixed a few issues from previous commit
Fixed an issue in the cube handler where a column identified as a
measure but not had the measure process completed doesn't come through
in the select and default view queries.

Added the ability to get cube views of specific revisions.

Exposed the full team details when requesting a datasets dto in all
languauges.

Added period coverage to the dataset as this is needed by the consumer
view.