Implement Reference Data
Continues the refactor of code to reduce duplicates.  Fixes a number of
bugs with lookup table and measure lookup table matching and makes sure
that all CSVs uploaded destined for DuckDB and now converted to UTF-8
before being passed through.

The key feature is the implementation of basic reference data with a
little bit of reference data validation which can be expanded on post
MVP.