Lay the foundations for the new update journey
This renames fact_tables to data_tables throughout the code base.  It
sets the relation of revisions to data_tables as a 1:1.  We're now
settled on the idea that one revision will only have one new data table.
Relocated the defination of the core fact table from revision 1 to a
dedicated "fact_table" defination which holds the columns which make up
the dataset and how they are defined.

Additional changes include cleaning up the routes so that routes are now
seperated out in to those which affect dataset, revision, dimension and
measure respectively.  This should hopefully make the code more
maintable and make it easier to track down issues.