Fixes long queries makes the cube builder a little more independant
The cube builder now only takes in IDs and is responsible for loading
its own dataset and associated relations.  Reduced the scope of the
default relations on the dataset.  This has cute queries from 15000ms
down to 800ms.

Confirmed cube builder and validation methods take around 500ms to
complete locally.   Looking to confirm in service to support our
rational for using DuckDB.