Fixes an issue where the cube bulder fails if select empty
This fixes an issue when the cube builder fails because there are no
dimensions or measures in a cube resulting in an empty select statement.
This add a check to see if there is something in the select and if not
adds `*` to the select to allow it just show the raw fact table.