Roll out languague matcher case statement to lookups
This was probably an oversight on my part but the languague algorithm I
created for measures should also be applied to lookup tables both in the
cube and in the during the journey.

The algorithm is now in its own function in the lookup-table-utils and
is used consistently where its needed.