Add downloading all cube data from the data lake as a zip
This improves raw file downloads for measure, dimension and data tables.
The first two were completely missing and are now needed now we don't
always have direcrt access to the datalake in the client estate.

Also added the option to pull down everything from the datalake for a
given dataset.  This is helpful when trying to understand why a given
dataset journey has failed.

Added a route to list all the file imports which made up a given
dataset.  Again this provides us with extra diagnositic information for
the developer page.

Fianlly fixed a small issue in the cube builder where there was a
slightly broken join statement on date dimensions as the date dimension
table switch from being monolinguagl to multilingual.