Fix for fact table column names with spaces
There were still a few places where column names weren't properly quoted
resulting in some SQL query failures.  This is the root cause of the bug
for SW 707.  Most people have been using StatsWales 2 formats but this
was the better StatsWales 3 format but its not been tested as strongly
as the other routes.