Align frontned views and start to add build log to cube handler
The Developer, Consumer and Preview views are now all using the same
view so it makes sense to align the views and remove line numbers.

Additionally spent some time adding additional logging and improving
some error handling in the cube handler.  Adding a build log we can pass
to the frontend to understand why a build has failed and at what point.