Complete the overhaul of the cube handler code
All the code which takes a user input in the form of column titles now
pass through the duckDBFormat function which currently wraps the
sqlstring's format function.

Next step is to switch out the duckDBFormat function and replace it with
the more appropriate @scaleleap/pg-format library.  This is geared
towards postgres syntax instead of mysql.  DuckDB and Postgres use
similar syntax.