Fix bugs around downloading and implement filtering and sorting
This PR aligns all the views to use the same code to query the postgres
cubes and implements interfaces for sorting and filtering the default
views.

Additionally all the download cube code has been ripped out so that it
no longer tries to generate the cube each time istead just going to the
datalake to get the required files.

Other fixes include the ignore column bug which is now fixed.  When
generating the cube all the files are generated at the same time
including the duckdb, csv, excel, json and parquet files.