Ensure unique column names in cube builder and fix sorting
Pair of things fixed here.  The first is column names need to be unique
in views.  While we can't stop a user using the same name for dimensions
and measures we can at least postfix them if we find duplicates.  This
prevents cube builds failing after poor test data is used in the
translation files.

The other thing fixed is natural languague sorting is mostly gone.
Sorting is now done against a sort column and then the columns natural
sort order.  This should result in more logical sorting of data.