Update error checking when alter table fails
Alter table throws a different error when there are null values in a
primary key column.  When we moved to postgres it was assumed the error
would still be the key creation failed but the error is actually
`contains null values`... Update the code to look for this error and run
the incomplete facts query.