Fix an issue in date matching where the year didn't advance
The display year was failing to advance when processing month data.
This will have caused issues for create and update journeies using month
based data.  The issue showed itself more when trying to create the
filter table in the cube builder as this failed due to a bad primary
key constraint.