SW-916 Make date matching BST aware
Date matching had largely assumed all dates were UTC/GMT but with the introduction of rolling years this stopped being the case.  So I've added the code to make the date matcher BST aware.  When it puts the date in the database its in UTC with the appropriate hour added if the date is in BST.