Improve code portability
Makes a number of changes to allow the code run on Windows, Linux and
MacOS.  First it removes the old multer config and directory creation
code.  Second it moves away from bash scripts in package.json when
trying to decide to run a command against podman or docker and instead
just uses the `docker-compose` command.  Additionally the duckdb tmp
directory has been updated to use `os.tmpdir()` instead of a hardcoded
`/tmp` which isn't available on Windows.