Add support for rolling years
Add support for rolling style periods.  Covers options for Year,
quarter, month and week ending periods.  Fixes issues with timezoned
dates being saved incorrectly to the database.

Also fixes a minor issue where when pulling the date out of the database
TypeOrm was trying to process the timezone information incorrectly
resulting in the wrong date/time going to the frontend.