Move from duckdb-async to duckdb-neo
DuckDB Async has been depricated in favour of duckdb-neo and is no
longer being updated.  For the long term support of the project its in
our interests to move from async to neo.

This change allowed us to go over the remaining duckdb calls and refine
the its usage slightly.  There's now a single duckdb instance created
and maintained in the app.  This allows us to allocate more resources to
duckdb while at the sametime controlling duckdbs use of available
container memory.  Another change is moving from duckdbs spacial plugin
for Excel handling to the newer built in Excel plugin.

Signed-off-by: Jamie Maynard <29251905+j-maynard@users.noreply.github.com>