Add better handling of missing codes and hierarchies for date dimension
So added the code to filter the resulting table for only values which
are in present in the input.  Also removed parents which aren't present
in the resulting looking data.

Additionally fixed the tests as the input needs to be complete
otherwise the resulting output is null.  Finally add the new code for
creating the lookup table in the database as in future PRs this code
will be removed from the cube builder and code using it pointed to this
function.