Refactor CSV PRocessor to incoming file processor
Refactors the CSV Processor into the Incoming File Processor which
mostly deals with ingesting a file using duckdb and returning a preview
of the file back.

Refactored out the code relating to getting a single column from the
fact table as this should live in the dimension processor as it deals
with getting a columns worth of data from a fact table.