Fixes an issue with builds when a revision is empty
The fix for incomplete incomplete builds broke times when revisions are
empty from a cube build point of view but we still need a cube based on
the details from the previous revision.

Added the build id to the returned object from the rebuild end points.

Finally fixed the SQL logic for checking for active bulk builds.