Adds a validation table to the cube
I've been toying for a while with doing validations against the filter
table to speed up queries against fact table columns when the fact
tables are really long.  While this seemed like a good idea the reality
is that the filter table reflects the output of the lookup tables not
the fact table.  So I've added a validation table as the first stage of
migrating queries to use this rather than the fact table.

Follow up work will switch column preview to use the output of the
validation table and to perform validations against this table.  After
that it'll be to use this table to clean up entries in the filter table
so it reflects only values present in the fact table rather than
everything in the lookup table.