Fixes an issue where errors querying the cube aren't handled
Fixes 2 issues.  The first is that when querying a bad cube the backend
fails to send any sort of response causing the frontend to just timeout
waiting for a response.

The second issue is we don't test the view before trying to materialise
it.  Added a line to test the first 500 records of a view to the cube
builder to catch view materialisation failures sooner.