Bootstrap creating validation table
For the cube bootstrap process to work there's an expectation that a
validation table exists in the cube already so that the date period
dimension lookup tables can be recreated.  Some of the older cubes in
production don't have this table present so the bootstrap and cube
rebuild process fails.  This should fix that issue.