Add coverage start and end dates to the revision
We have some logic on the dataset dto to pull these from the latest
published revision but this logic can only trigger if there's a
published revision in the requested object.

Adding it to the revision means it'll be there when the revisions
metadata is pulled down.  It also means we can check to make sure the
correct dates are being populated.